Gradient perturbation to create numeric Hessian
Attempt at creating a numeric Hessian from a gradient perturbation. The
idea (from Tore) was quite interesting, and seems to work for some of
the elements - that is, give the same result as the analytic Hessian.

However, many of the elements are not the same - and it is not
completely clear whether this is due to actual errors in the analytic
Hessian or whether it is due to a mistake in the numeric approximation.