Error hunting in the Hessian
The whole day - give or take - has been used to hunt for errors in the
Hessian formulation. Using central differences - after tip from Tore -
only dmu/dn still holds an error.

Mu in itself seems to be correct - perturbation in the mole numbers on
A confirms. Thus the error seems to be in the derivative itself. Caught
a stupid error in the ideal gas Hessian, so it is (almost) right. Seems
to be on the same order of magnitude as Kasper and Fredrik, but no
convergence in sight...

I will have to continue the hunt tomorrow; not sure what I'll do if the
mistake cannot be found.