Checking results
The issue with the analytic Hessian is still not resolved, and I'm
still using the numeric Hessian. This seems to work satisfactory.

Units - or rather - the scaling of units seems to somehow have
introduced a problem of unknown origin. Changing back to SI units
seemed to help. It is hard to explain why this seems to be the case.